Conform names to spreadsheet workbook standards

scrub_tabnames() preserves as much of the original tabnames as possible while ensuring each tabname is unique and contains only characters permissible as spreadsheet tab names

Usage

scrub_tabnames(
  tabnames,
  max_width = 31,
  truncate_side = c("right", "left", "center"),
  paste_side = c("right", "left"),
  sep = ".",
  pad = ".",
  quiet = FALSE
)

Arguments

tabnames

strings: character vector of names

max_width

integerish between 0 and 31: Default is 31 characters; longer names will cause Excel export to fail.

truncate_side

'right', 'left', or 'center': describes which side of a long tabname do you want to snip excess characters from

paste_side

'right' or 'left': describing which side of a tab name you want to paste characters to, if required

sep

chars: if pasting characters to a tab name, what character do you want to use to separate name and unique prefix/suffix; can be several characters like '...', but cannot be any of the forbidden characters described in details

pad

char: if pasting characters to a tab name, what character do you want to use to pad so numbers align, (e.g. '0' for '001' or '.' for '..1' if 3 digits of differentiation are necessary). Must be at least 1 character but no longer. Can be a space ' ', but cannot be any of the forbidden characters described in details

quiet

bool: to turn off warnings if you prefer with quiet = TRUE

Value

a character vector of names suitable for Excel or Librecalc tab/sheet names

Details

Replaces:

characters: /\: with: -

characters: ?*&$ with: #

characters: [< with: (

characters: ]> with: )

tabnames with the single quote or double quote: ' or " with a backtick `

the word 'history' in a tabname, with 'hist' while preserving capitalization

names greater than 31 characters with a truncated characters

duplicate sheetnames with numbered versions of the name to make each name within that group unique

References

https://help.libreoffice.org/latest/he/text/scalc/guide/rename_table.html

https://support.microsoft.com/en-us/office/rename-a-worksheet-3f1f7148-ee83-404d-8ef0-9ff99fbad1f9

https://colinfay.me/writing-r-extensions/creating-r-packages.html

Examples

scrub_tabnames(1:5)
#> [1] "1" "2" "3" "4" "5"
c('\\:blue/:', 'red', 'gr?*een///////', '[]', '[orange]','', NA) |> scrub_tabnames()
#> ! Sheet names should not contain:  `\`
#> → Replacing forbidden character with:  `-`
#> ! Sheet names should not contain:  `/`
#> → Replacing forbidden character with:  `-`
#> ! Sheet names should not contain:  `:`
#> → Replacing forbidden character with:  `-`
#> ! Sheet names should not contain:  `*`
#> → Replacing forbidden character with:  `#`
#> ! Sheet names should not contain:  `?`
#> → Replacing forbidden character with:  `#`
#> ! Sheet names should not contain:  `[`
#> → Replacing forbidden character with:  `(`
#> ! Sheet names should not contain:  `]`
#> → Replacing forbidden character with:  `)`
#> ! Duplicate or empty tab/sheet names exist. Numbering within each group ... 
#> [1] "--blue--"       "red"            "gr##een-------" "()"            
#> [5] "(orange)"       ".1"             "NA"            
c('history', NA, 'HISTORY', 'Entire History of the Universe',NULL,NULL) |> scrub_tabnames()
#> ! Sheet names should not contain:  `the word 'history'`
#> → Replacing forbidden character with:  `hist`
#> ! Duplicate or empty tab/sheet names exist. Numbering within each group ... 
#> [1] "hist.1"                      "NA"                         
#> [3] "HIST.2"                      "Entire Hist of the Universe"
c("'\'don't worry\''", '`NA`', NA, "'HIS'T''", 'Entire History of \'the\' \'Universe\'') |> scrub_tabnames()
#> ! Sheet names should not contain:  `a single quote ' at the start of a tabname`
#> → Replacing forbidden character with:  `an empty string`
#> ! Sheet names should not contain:  `a single quote ' at the end of a tabname`
#> → Replacing forbidden character with:  `an empty string`
#> ! Sheet names should not contain:  `the word 'history'`
#> → Replacing forbidden character with:  `hist`
#> [1] "`don't worry`"                "`NA`"                        
#> [3] "NA"                           "`HIS'T`"                     
#> [5] "Entire Hist of 'the' 'Univer"
c('\\/:[]?*', '\\?:*/[]', '~!@#$%^&()-_=+{}|;:,<.> ') |> scrub_tabnames()
#> ! Sheet names should not contain:  `\`
#> → Replacing forbidden character with:  `-`
#> ! Sheet names should not contain:  `/`
#> → Replacing forbidden character with:  `-`
#> ! Sheet names should not contain:  `:`
#> → Replacing forbidden character with:  `-`
#> ! Sheet names should not contain:  `*`
#> → Replacing forbidden character with:  `#`
#> ! Sheet names should not contain:  `?`
#> → Replacing forbidden character with:  `#`
#> ! Sheet names should not contain:  `?`
#> → Replacing forbidden character with:  `#`
#> ! Sheet names should not contain:  `[`
#> → Replacing forbidden character with:  `(`
#> ! Sheet names should not contain:  `]`
#> → Replacing forbidden character with:  `)`
#> ! Sheet names should not contain:  `<`
#> → Replacing forbidden character with:  `(`
#> ! Sheet names should not contain:  `>`
#> → Replacing forbidden character with:  `)`
#> [1] "---()##"                  "-#-#-()"                 
#> [3] "~!@#$%^#()-_=+{}|;-,(.) "
paste0(names(datasets::precip),'/',datasets::precip) |> scrub_tabnames(max_width = 10)
#> ! Sheet names should not contain:  `/`
#> → Replacing forbidden character with:  `-`
#>  [1] "Mobile-67"  "Juneau-54." "Phoenix-7"  "Little Roc" "Los Angele"
#>  [6] "Sacramento" "San Franci" "Denver-13"  "Hartford-4" "Wilmington"
#> [11] "Washington" "Jacksonvil" "Miami-59.8" "Atlanta-48" "Honolulu-2"
#> [16] "Boise-11.5" "Chicago-34" "Peoria-35." "Indianapol" "Des Moines"
#> [21] "Wichita-30" "Louisville" "New Orlean" "Portland-4" "Baltimore-"
#> [26] "Boston-42." "Detroit-31" "Sault Ste." "Duluth-30." "Minneapoli"
#> [31] "Jackson-49" "Kansas Cit" "St Louis-3" "Great Fall" "Omaha-30.2"
#> [36] "Reno-7.2"   "Concord-36" "Atlantic C" "Albuquerqu" "Albany-33."
#> [41] "Buffalo-36" "New York-4" "Charlotte-" "Raleigh-42" "Bismark-16"
#> [46] "Cincinnati" "Cleveland-" "Columbus-3" "Oklahoma C" "Portland-3"
#> [51] "Philadelph" "Pittsburg-" "Providence" "Columbia-4" "Sioux Fall"
#> [56] "Memphis-49" "Nashville-" "Dallas-35." "El Paso-7." "Houston-48"
#> [61] "Salt Lake " "Burlington" "Norfolk-44" "Richmond-4" "Seattle Ta"
#> [66] "Spokane-17" "Charleston" "Milwaukee-" "Cheyenne-1" "San Juan-5"
paste0(names(datasets::precip),'/',datasets::precip) |> scrub_tabnames(max_width = 8, sep = '_#', pad = '0')
#> ! Sheet names should not contain:  `/`
#> → Replacing forbidden character with:  `-`
#> ! Duplicate or empty tab/sheet names exist. Numbering within each group ... 
#>  [1] "Mobile-6" "Juneau-5" "Phoenix-" "Little R" "Los Ange" "Sacramen"
#>  [7] "San Fran" "Denver-1" "Hartford" "Wilmingt" "Washingt" "Jacksonv"
#> [13] "Miami-59" "Atlanta-" "Honolulu" "Boise-11" "Chicago-" "Peoria-3"
#> [19] "Indianap" "Des Moin" "Wichita-" "Louisvil" "New Orle" "Portl_#1"
#> [25] "Baltimor" "Boston-4" "Detroit-" "Sault St" "Duluth-3" "Minneapo"
#> [31] "Jackson-" "Kansas C" "St Louis" "Great Fa" "Omaha-30" "Reno-7.2"
#> [37] "Concord-" "Atlantic" "Albuquer" "Albany-3" "Buffalo-" "New York"
#> [43] "Charlott" "Raleigh-" "Bismark-" "Cincinna" "Clevelan" "Columbus"
#> [49] "Oklahoma" "Portl_#2" "Philadel" "Pittsbur" "Providen" "Columbia"
#> [55] "Sioux Fa" "Memphis-" "Nashvill" "Dallas-3" "El Paso-" "Houston-"
#> [61] "Salt Lak" "Burlingt" "Norfolk-" "Richmond" "Seattle " "Spokane-"
#> [67] "Charlest" "Milwauke" "Cheyenne" "San Juan"
paste0(names(datasets::precip),'/',datasets::precip) |> scrub_tabnames(max_width = 0, sep = '...', pad = '0')
#> ! The minimum width must be greater than the number of characters required
#>         to differentiate possible duplicated tabnames plus the characters in `sep`.
#>         
#> Setting the tabname character width to 5.
#> ! Sheet names should not contain:  `/`
#> → Replacing forbidden character with:  `-`
#> ! Duplicate or empty tab/sheet names exist. Numbering within each group ... 
#> ! Duplicates remain after replacements. Trying again ... 
#> ! Duplicate or empty tab/sheet names exist. Numbering within each group ... 
#>  [1] "Mobil" "Junea" "Phoen" "Littl" "Los A" "Sacra" "San F" "Denve" "Hartf"
#> [10] "Wilmi" "Washi" "J...1" "Miami" "A...1" "Honol" "Boise" "Chica" "Peori"
#> [19] "India" "Des M" "Wichi" "Louis" "New O" "P...1" "Balti" "Bosto" "Detro"
#> [28] "Sault" "Dulut" "Minne" "J...2" "Kansa" "St Lo" "Great" "Omaha" "Reno-"
#> [37] "Conco" "A...2" "Albuq" "Alban" "Buffa" "New Y" "C...1" "Ralei" "Bisma"
#> [46] "Cinci" "Cleve" "C...2" "Oklah" "P...2" "Phila" "Pitts" "Provi" "C...3"
#> [55] "Sioux" "Memph" "Nashv" "Dalla" "El Pa" "Houst" "Salt " "Burli" "Norfo"
#> [64] "Richm" "Seatt" "Spoka" "C...4" "Milwa" "Cheye" "San J"
paste0(names(datasets::precip),'/',datasets::precip) |> scrub_tabnames(max_width = 0, sep = '..', pad = '.')
#> ! The minimum width must be greater than the number of characters required
#>         to differentiate possible duplicated tabnames plus the characters in `sep`.
#>         
#> Setting the tabname character width to 4.
#> ! Duplicate or empty tab/sheet names exist. Numbering within each group ... 
#> ! Duplicates remain after replacements. Trying again ... 
#> ! Duplicate or empty tab/sheet names exist. Numbering within each group ... 
#>  [1] "Mobi" "June" "Phoe" "Litt" "Los " "Sacr" "S..1" "Denv" "Hart" "Wilm"
#> [11] "Wash" "J..1" "Miam" "A..1" "Hono" "Bois" "Chic" "Peor" "Indi" "Des "
#> [21] "Wich" "Loui" "N..1" "P..1" "Balt" "Bost" "Detr" "Saul" "Dulu" "Minn"
#> [31] "J..2" "Kans" "St L" "Grea" "Omah" "Reno" "Conc" "A..2" "Albu" "Alba"
#> [41] "Buff" "N..2" "C..1" "Rale" "Bism" "Cinc" "Clev" "C..2" "Okla" "P..2"
#> [51] "Phil" "Pitt" "Prov" "C..3" "Siou" "Memp" "Nash" "Dall" "El P" "Hous"
#> [61] "Salt" "Burl" "Norf" "Rich" "Seat" "Spok" "C..4" "Milw" "Chey" "S..2"
rep('', 15) |> scrub_tabnames(max_width = 0, sep = '..', pad = '.')
#> ! The minimum width must be greater than the number of characters required
#>         to differentiate possible duplicated tabnames plus the characters in `sep`.
#>         
#> Setting the tabname character width to 4.
#> ! Duplicate or empty tab/sheet names exist. Numbering within each group ... 
#>  [1] "...1" "...2" "...3" "...4" "...5" "...6" "...7" "...8" "...9" "..10"
#> [11] "..11" "..12" "..13" "..14" "..15"
# rep('', 15) |> scrub_tabnames(max_width = 0, sep = '', pad = '') # expect_fail
# rep('', 15) |> scrub_tabnames(max_width = 0, sep = '', pad = ' ') # expect_fail
# dplyr::starwars |> dplyr::mutate(new_name = paste(name,'of', homeworld)) |>
# dplyr::pull(new_name) |> scrub_tabnames(max_width = 20, sep = '//', pad = ']') # expect_fail
# dplyr::starwars |> dplyr::mutate(new_name = paste(name,'of', homeworld)) |>
# dplyr::pull(new_name) |> scrub_tabnames(max_width = 0, sep = '.', pad = '.') # expect_fail

# when stringr::str_trunc() gets fixed, the below code will work.
# dplyr::starwars |> dplyr::mutate(new_name = paste(name,'of', homeworld)) |>
# dplyr::pull(new_name) |> scrub_tabnames(max_width = 0, sep = '', pad = '.') # ??????

# if you request a width of characters that is fewer than your replacement characters, you get some weird looking names:
dplyr::starwars |> dplyr::mutate(new_name = paste(name,'of', homeworld)) |>
dplyr::pull(new_name) |> scrub_tabnames(max_width = 2, sep = '...', pad = '.')
#> ! The minimum width must be greater than the number of characters required
#>         to differentiate possible duplicated tabnames plus the characters in `sep`.
#>         
#> Setting the tabname character width to 5.
#> ! Duplicate or empty tab/sheet names exist. Numbering within each group ... 
#>  [1] "Luke " "C-3PO" "R2-D2" "D...1" "Leia " "Owen " "Beru " "R5-D4" "Biggs"
#> [10] "Obi-W" "Anaki" "Wilhu" "Chewb" "Han S" "Greed" "Jabba" "Wedge" "Jek T"
#> [19] "Yoda " "Palpa" "Boba " "IG-88" "Bossk" "Lando" "Lobot" "Ackba" "Mon M"
#> [28] "Arvel" "Wicke" "Nien " "Qui-G" "Nute " "Finis" "Jar J" "Roos " "Rugor"
#> [37] "Ric O" "Watto" "Sebul" "Quars" "Shmi " "D...2" "Bib F" "Ayla " "Dud B"
#> [46] "Gasga" "Ben Q" "Mace " "Ki-Ad" "Kit F" "Eeth " "Adi G" "Saese" "Yarae"
#> [55] "Plo K" "Mas A" "Grega" "Cordé" "Clieg" "Poggl" "Lumin" "Barri" "Dormé"
#> [64] "Dooku" "Bail " "Jango" "Zam W" "Dexte" "Lama " "Taun " "Jocas" "Ratts"
#> [73] "R4-P1" "Wat T" "San H" "Shaak" "Griev" "Tarff" "Raymu" "Sly M" "Tion "
#> [82] "Finn " "Rey o" "Poe D" "BB8 o" "Capta" "Padmé"

# However, if you call zero width..., that's useful for naming.
dplyr::starwars |> dplyr::mutate(new_name = paste(name,'of', homeworld)) |>
dplyr::pull(new_name) |> scrub_tabnames(max_width = 0)
#> ! The minimum width must be greater than the number of characters required
#>         to differentiate possible duplicated tabnames plus the characters in `sep`.
#>         
#> Setting the tabname character width to 3.
#> ! Duplicate or empty tab/sheet names exist. Numbering within each group ... 
#>  [1] "Luk" "C-3" "R2-" "D.1" "Lei" "Owe" "Ber" "R5-" "Big" "Obi" "Ana" "Wil"
#> [13] "Che" "Han" "G.1" "Jab" "Wed" "Jek" "Yod" "Pal" "Bob" "IG-" "Bos" "Lan"
#> [25] "Lob" "Ack" "Mon" "Arv" "Wic" "Nie" "Qui" "Nut" "F.1" "Jar" "Roo" "Rug"
#> [37] "Ric" "W.1" "Seb" "Qua" "Shm" "D.2" "Bib" "Ayl" "Dud" "Gas" "Ben" "Mac"
#> [49] "Ki-" "Kit" "Eet" "Adi" "Sae" "Yar" "Plo" "Mas" "G.2" "Cor" "Cli" "Pog"
#> [61] "Lum" "Bar" "Dor" "Doo" "Bai" "Jan" "Zam" "Dex" "Lam" "Tau" "Joc" "Rat"
#> [73] "R4-" "W.2" "San" "Sha" "Gri" "Tar" "Ray" "Sly" "Tio" "F.2" "Rey" "Poe"
#> [85] "BB8" "Cap" "Pad"
dplyr::starwars |> dplyr::mutate(new_name = paste(name,'of', homeworld)) |>
dplyr::pull(new_name) |> scrub_tabnames(max_width = 0, sep = 'tab > ', pad = '0')
#> ! Sheet names should not contain:  `>`
#> → Replacing forbidden character with:  `)`
#> ! The minimum width must be greater than the number of characters required
#>         to differentiate possible duplicated tabnames plus the characters in `sep`.
#>         
#> Setting the tabname character width to 8.
#>  [1] "Luke Sky" "C-3PO of" "R2-D2 of" "Darth Va" "Leia Org" "Owen Lar"
#>  [7] "Beru Whi" "R5-D4 of" "Biggs Da" "Obi-Wan " "Anakin S" "Wilhuff "
#> [13] "Chewbacc" "Han Solo" "Greedo o" "Jabba De" "Wedge An" "Jek Tono"
#> [19] "Yoda of " "Palpatin" "Boba Fet" "IG-88 of" "Bossk of" "Lando Ca"
#> [25] "Lobot of" "Ackbar o" "Mon Moth" "Arvel Cr" "Wicket S" "Nien Nun"
#> [31] "Qui-Gon " "Nute Gun" "Finis Va" "Jar Jar " "Roos Tar" "Rugor Na"
#> [37] "Ric Olié" "Watto of" "Sebulba " "Quarsh P" "Shmi Sky" "Darth Ma"
#> [43] "Bib Fort" "Ayla Sec" "Dud Bolt" "Gasgano " "Ben Quad" "Mace Win"
#> [49] "Ki-Adi-M" "Kit Fist" "Eeth Kot" "Adi Gall" "Saesee T" "Yarael P"
#> [55] "Plo Koon" "Mas Amed" "Gregar T" "Cordé of" "Cliegg L" "Poggle t"
#> [61] "Luminara" "Barriss " "Dormé of" "Dooku of" "Bail Pre" "Jango Fe"
#> [67] "Zam Wese" "Dexter J" "Lama Su " "Taun We " "Jocasta " "Ratts Ty"
#> [73] "R4-P17 o" "Wat Tamb" "San Hill" "Shaak Ti" "Grievous" "Tarfful "
#> [79] "Raymus A" "Sly Moor" "Tion Med" "Finn of " "Rey of N" "Poe Dame"
#> [85] "BB8 of N" "Captain " "Padmé Am"
dplyr::starwars |> dplyr::mutate(new_name = paste(name,'of', homeworld)) |>
dplyr::pull(new_name) |> scrub_tabnames(max_width = 0,truncate_side = 'center')
#> ! The minimum width must be greater than the number of characters required
#>         to differentiate possible duplicated tabnames plus the characters in `sep`.
#>         
#> Setting the tabname character width to 3.
#> ! Duplicate or empty tab/sheet names exist. Numbering within each group ... 
#> 
#> ****************************************************************************
#> ***** There are still issues after scrubbing names. Rerunning function *****
#> ****************************************************************************
#> 
#>  [1] "Lue" "C-e" "R2o" "Dae" "Len" "Owe" "Bee" "R5e" "Bie" "Obn" "Ane" "Wiu"
#> [13] "Chk" "Haa" "Gra" "Jaa" "Wea" "JeV" "YoA" "P~1" "Boo" "IGA" "Boa" "L~1"
#> [25] "Lon" "Aca" "Moa" "ArA" "Wir" "Nit" "QuA" "Nua" "Fit" "Jao" "Roo" "Ruo"
#> [37] "Rio" "Waa" "See" "Quo" "She" "Dar" "Bih" "Ayh" "Dur" "Gan" "Bed" "Mal"
#> [49] "Kia" "Kim" "Eea" "Adt" "Sah" "Yaa" "Pln" "Maa" "Gro" "Coo" "Cle" "Pos"
#> [61] "Lul" "Bal" "D~1" "D~2" "Ban" "Jan" "Zan" "Dem" "L~2" "Tao" "Jot" "Rar"
#> [73] "R4A" "Wao" "Sat" "Shi" "Gre" "Tak" "Ran" "Sla" "Tiu" "FiA" "ReA" "PoA"
#> [85] "BBA" "CaA" "P~2"