Convert a data object or list of data objects to a tibble with rownames

Same as tibble() with different defaults. entibble:

Always includes rownames, if available, as the first column of the tibble

Is not concerned with duplicate column names unless specified by user with .name_repair

Does not repeat input elements to obtain common lengths

Transforms lists with uniform arrays into columns of a tibble

Treats ragged list of input elements similar to tibble() but includes names as column if possible

Arranges lists of lists to produce a single table, if row dimensions allow it

Usage

entibble(
  ...,
  .rowname = "rowname",
  .name_repair = c("minimal", "check_unique", "unique", "universal", "unique_quiet",
    "universal_quiet")
)

Arguments

...

object or expression to convert to a tibble

.rowname

string; to name the column containing rownames. If rownames are not present in the data, .rowname is ignored.

.name_repair

Treatment of problematic column names:

"minimal": No name repair or checks, beyond basic existence,

"unique": Make sure names are unique and not empty,

"check_unique": (default value), no name repair, but check they are unique,

"universal": Make the names unique and syntactic

a function: apply custom name repair (e.g., .name_repair = make.names for names in the style of base R).

A purrr-style anonymous function, see rlang::as_function()

This argument is passed on as repair to vctrs::vec_as_names(). See there for more details on these terms and the strategies used to enforce them.

Value

a tibble with rownames in a column

Details

@details Entibble is designed to work with the function xl() to produce a viewable output of R data in spreadsheets. To help xl() succeed as often as possible, entibble() is liberal about names, such as duplicate column names, and makes an effort to produce a flat table in place of a nested tibble.

See also

Examples

#TODO: transport these to tests and decide what needs to stay...
entibble('')
#> # A tibble: 1 × 1
#>   ``   
#>   <chr>
#> 1 ""   
letters |> entibble()
#> # A tibble: 26 × 1
#>    letters
#>    <chr>  
#>  1 a      
#>  2 b      
#>  3 c      
#>  4 d      
#>  5 e      
#>  6 f      
#>  7 g      
#>  8 h      
#>  9 i      
#> 10 j      
#> # ℹ 16 more rows
letters |>  purrr::set_names(LETTERS) |> entibble()
#> # A tibble: 26 × 2
#>    rowname `purrr::set_names(letters, LETTERS)`
#>    <chr>   <chr>                               
#>  1 A       a                                   
#>  2 B       b                                   
#>  3 C       c                                   
#>  4 D       d                                   
#>  5 E       e                                   
#>  6 F       f                                   
#>  7 G       g                                   
#>  8 H       h                                   
#>  9 I       i                                   
#> 10 J       j                                   
#> # ℹ 16 more rows
letters |>  purrr::set_names(LETTERS) |> entibble(.rowname = 'user-specified rowname')
#> # A tibble: 26 × 2
#>    `user-specified rowname` `purrr::set_names(letters, LETTERS)`
#>    <chr>                    <chr>                               
#>  1 A                        a                                   
#>  2 B                        b                                   
#>  3 C                        c                                   
#>  4 D                        d                                   
#>  5 E                        e                                   
#>  6 F                        f                                   
#>  7 G                        g                                   
#>  8 H                        h                                   
#>  9 I                        i                                   
#> 10 J                        j                                   
#> # ℹ 16 more rows
(example_data <- letters |>  purrr::set_names(LETTERS))
#>   A   B   C   D   E   F   G   H   I   J   K   L   M   N   O   P   Q   R   S   T 
#> "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s" "t" 
#>   U   V   W   X   Y   Z 
#> "u" "v" "w" "x" "y" "z" 
example_data |> entibble()
#> # A tibble: 26 × 2
#>    rowname example_data
#>    <chr>   <chr>       
#>  1 A       a           
#>  2 B       b           
#>  3 C       c           
#>  4 D       d           
#>  5 E       e           
#>  6 F       f           
#>  7 G       g           
#>  8 H       h           
#>  9 I       i           
#> 10 J       j           
#> # ℹ 16 more rows
example_data |> entibble(.rowname = 'user-specified rowname')
#> # A tibble: 26 × 2
#>    `user-specified rowname` example_data
#>    <chr>                    <chr>       
#>  1 A                        a           
#>  2 B                        b           
#>  3 C                        c           
#>  4 D                        d           
#>  5 E                        e           
#>  6 F                        f           
#>  7 G                        g           
#>  8 H                        h           
#>  9 I                        i           
#> 10 J                        j           
#> # ℹ 16 more rows
example_data |> entibble(.name_repair = ~c('name_A', 'name_B'))
#> # A tibble: 26 × 2
#>    name_A name_B
#>    <chr>  <chr> 
#>  1 A      a     
#>  2 B      b     
#>  3 C      c     
#>  4 D      d     
#>  5 E      e     
#>  6 F      f     
#>  7 G      g     
#>  8 H      h     
#>  9 I      i     
#> 10 J      j     
#> # ℹ 16 more rows
tibble::tibble(example_data, !!!letters) # ragged data
#> # A tibble: 26 × 27
#>    example_data `"a"` `"b"` `"c"` `"d"` `"e"` `"f"` `"g"` `"h"` `"i"` `"j"`
#>    <chr>        <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#>  1 a            a     b     c     d     e     f     g     h     i     j    
#>  2 b            a     b     c     d     e     f     g     h     i     j    
#>  3 c            a     b     c     d     e     f     g     h     i     j    
#>  4 d            a     b     c     d     e     f     g     h     i     j    
#>  5 e            a     b     c     d     e     f     g     h     i     j    
#>  6 f            a     b     c     d     e     f     g     h     i     j    
#>  7 g            a     b     c     d     e     f     g     h     i     j    
#>  8 h            a     b     c     d     e     f     g     h     i     j    
#>  9 i            a     b     c     d     e     f     g     h     i     j    
#> 10 j            a     b     c     d     e     f     g     h     i     j    
#> # ℹ 16 more rows
#> # ℹ 16 more variables: `"k"` <chr>, `"l"` <chr>, `"m"` <chr>, `"n"` <chr>,
#> #   `"o"` <chr>, `"p"` <chr>, `"q"` <chr>, `"r"` <chr>, `"s"` <chr>,
#> #   `"t"` <chr>, `"u"` <chr>, `"v"` <chr>, `"w"` <chr>, `"x"` <chr>,
#> #   `"y"` <chr>, `"z"` <chr>
entibble(example_data, !!!letters) # ragged data
#> # A tibble: 27 × 2
#>    rowname      `list(ragged_elements)`
#>    <chr>        <list>                 
#>  1 example_data <tibble [26 × 2]>      
#>  2 a            <tibble [1 × 1]>       
#>  3 b            <tibble [1 × 1]>       
#>  4 c            <tibble [1 × 1]>       
#>  5 d            <tibble [1 × 1]>       
#>  6 e            <tibble [1 × 1]>       
#>  7 f            <tibble [1 × 1]>       
#>  8 g            <tibble [1 × 1]>       
#>  9 h            <tibble [1 × 1]>       
#> 10 i            <tibble [1 × 1]>       
#> # ℹ 17 more rows
entibble(!!!letters)
#> # A tibble: 1 × 26
#>   a     b     c     d     e     f     g     h     i     j     k     l     m    
#>   <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#> 1 a     b     c     d     e     f     g     h     i     j     k     l     m    
#> # ℹ 13 more variables: n <chr>, o <chr>, p <chr>, q <chr>, r <chr>, s <chr>,
#> #   t <chr>, u <chr>, v <chr>, w <chr>, x <chr>, y <chr>, z <chr>
tibble::tibble(!!!letters)
#> # A tibble: 1 × 26
#>   `"a"` `"b"` `"c"` `"d"` `"e"` `"f"` `"g"` `"h"` `"i"` `"j"` `"k"` `"l"` `"m"`
#>   <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#> 1 a     b     c     d     e     f     g     h     i     j     k     l     m    
#> # ℹ 13 more variables: `"n"` <chr>, `"o"` <chr>, `"p"` <chr>, `"q"` <chr>,
#> #   `"r"` <chr>, `"s"` <chr>, `"t"` <chr>, `"u"` <chr>, `"v"` <chr>,
#> #   `"w"` <chr>, `"x"` <chr>, `"y"` <chr>, `"z"` <chr>
entibble(example_data, example_data)
#> # A tibble: 26 × 4
#>    rowname example_data rowname example_data
#>    <chr>   <chr>        <chr>   <chr>       
#>  1 A       a            A       a           
#>  2 B       b            B       b           
#>  3 C       c            C       c           
#>  4 D       d            D       d           
#>  5 E       e            E       e           
#>  6 F       f            F       f           
#>  7 G       g            G       g           
#>  8 H       h            H       h           
#>  9 I       i            I       i           
#> 10 J       j            J       j           
#> # ℹ 16 more rows
# tibble::tibble(example_data, example_data)

letters |> purrr::set_names(LETTERS) |> as.list() |> tibble::tibble()
#> # A tibble: 26 × 1
#>    `as.list(purrr::set_names(letters, LETTERS))`
#>    <named list>                                 
#>  1 <chr [1]>                                    
#>  2 <chr [1]>                                    
#>  3 <chr [1]>                                    
#>  4 <chr [1]>                                    
#>  5 <chr [1]>                                    
#>  6 <chr [1]>                                    
#>  7 <chr [1]>                                    
#>  8 <chr [1]>                                    
#>  9 <chr [1]>                                    
#> 10 <chr [1]>                                    
#> # ℹ 16 more rows
letters |> purrr::set_names(LETTERS) |> as.list() |> entibble()
#> # A tibble: 1 × 26
#>   A     B     C     D     E     F     G     H     I     J     K     L     M    
#>   <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#> 1 a     b     c     d     e     f     g     h     i     j     k     l     m    
#> # ℹ 13 more variables: N <chr>, O <chr>, P <chr>, Q <chr>, R <chr>, S <chr>,
#> #   T <chr>, U <chr>, V <chr>, W <chr>, X <chr>, Y <chr>, Z <chr>

enlist(head(iris), tail(mtcars)) |> tibble::tibble()
#> # A tibble: 2 × 1
#>   `enlist(head(iris), tail(mtcars))`
#>   <named list>                      
#> 1 <df [6 × 5]>                      
#> 2 <df [6 × 11]>                     
enlist(head(iris), tail(mtcars)) |> entibble()
#> # A tibble: 6 × 17
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species rowname    mpg   cyl
#>   <chr>        <chr>       <chr>        <chr>       <chr>   <chr>    <dbl> <dbl>
#> 1 5.1          3.5         1.4          0.2         setosa  Porsche…  26       4
#> 2 4.9          3.0         1.4          0.2         setosa  Lotus E…  30.4     4
#> 3 4.7          3.2         1.3          0.2         setosa  Ford Pa…  15.8     8
#> 4 4.6          3.1         1.5          0.2         setosa  Ferrari…  19.7     6
#> 5 5.0          3.6         1.4          0.2         setosa  Maserat…  15       8
#> 6 5.4          3.9         1.7          0.4         setosa  Volvo 1…  21.4     4
#> # ℹ 9 more variables: disp <dbl>, hp <dbl>, drat <dbl>, wt <dbl>, qsec <dbl>,
#> #   vs <dbl>, am <dbl>, gear <dbl>, carb <dbl>

enlist(head(iris), head(mtcars)) |> entibble()
#> # A tibble: 6 × 17
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species rowname    mpg   cyl
#>   <chr>        <chr>       <chr>        <chr>       <chr>   <chr>    <dbl> <dbl>
#> 1 5.1          3.5         1.4          0.2         setosa  Mazda R…  21       6
#> 2 4.9          3.0         1.4          0.2         setosa  Mazda R…  21       6
#> 3 4.7          3.2         1.3          0.2         setosa  Datsun …  22.8     4
#> 4 4.6          3.1         1.5          0.2         setosa  Hornet …  21.4     6
#> 5 5.0          3.6         1.4          0.2         setosa  Hornet …  18.7     8
#> 6 5.4          3.9         1.7          0.4         setosa  Valiant   18.1     6
#> # ℹ 9 more variables: disp <dbl>, hp <dbl>, drat <dbl>, wt <dbl>, qsec <dbl>,
#> #   vs <dbl>, am <dbl>, gear <dbl>, carb <dbl>
entibble(head(iris), head(mtcars))
#> # A tibble: 6 × 17
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species rowname    mpg   cyl
#>   <chr>        <chr>       <chr>        <chr>       <chr>   <chr>    <dbl> <dbl>
#> 1 5.1          3.5         1.4          0.2         setosa  Mazda R…  21       6
#> 2 4.9          3.0         1.4          0.2         setosa  Mazda R…  21       6
#> 3 4.7          3.2         1.3          0.2         setosa  Datsun …  22.8     4
#> 4 4.6          3.1         1.5          0.2         setosa  Hornet …  21.4     6
#> 5 5.0          3.6         1.4          0.2         setosa  Hornet …  18.7     8
#> 6 5.4          3.9         1.7          0.4         setosa  Valiant   18.1     6
#> # ℹ 9 more variables: disp <dbl>, hp <dbl>, drat <dbl>, wt <dbl>, qsec <dbl>,
#> #   vs <dbl>, am <dbl>, gear <dbl>, carb <dbl>
tibble::tibble(head(iris), head(mtcars))
#> # A tibble: 6 × 16
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species   mpg   cyl  disp
#>          <dbl>       <dbl>        <dbl>       <dbl> <fct>   <dbl> <dbl> <dbl>
#> 1          5.1         3.5          1.4         0.2 setosa   21       6   160
#> 2          4.9         3            1.4         0.2 setosa   21       6   160
#> 3          4.7         3.2          1.3         0.2 setosa   22.8     4   108
#> 4          4.6         3.1          1.5         0.2 setosa   21.4     6   258
#> 5          5           3.6          1.4         0.2 setosa   18.7     8   360
#> 6          5.4         3.9          1.7         0.4 setosa   18.1     6   225
#> # ℹ 8 more variables: hp <dbl>, drat <dbl>, wt <dbl>, qsec <dbl>, vs <dbl>,
#> #   am <dbl>, gear <dbl>, carb <dbl>
entibble(head(iris), head(mtcars, 11))
#> # A tibble: 2 × 2
#>   rowname          `list(ragged_elements)`
#>   <chr>            <list>                 
#> 1 head(iris)       <tibble [6 × 5]>       
#> 2 head(mtcars, 11) <tibble [11 × 12]>     
# tibble::tibble(head(iris), head(mtcars,11))

tibble::tibble(y = 1:3,x = 11:13)
#> # A tibble: 3 × 2
#>       y     x
#>   <int> <int>
#> 1     1    11
#> 2     2    12
#> 3     3    13
entibble(y = c(1, 2, 3),x = c(11, 12, 13))
#> # A tibble: 3 × 2
#>       y     x
#>   <dbl> <dbl>
#> 1     1    11
#> 2     2    12
#> 3     3    13
# tibble::tibble(x = 1:3,x = 11:13)
entibble(x = 1:3, x = 11:13)
#> # A tibble: 3 × 2
#>       x     x
#>   <int> <int>
#> 1     1    11
#> 2     2    12
#> 3     3    13
entibble(x = 1:3,x = 10:14, .name_repair = 'minimal')
#> # A tibble: 2 × 2
#>   rowname `list(ragged_elements)`
#>   <chr>   <list>                 
#> 1 x       <tibble [3 × 1]>       
#> 2 x       <tibble [5 × 1]>       
# tibble::tibble(x = 1:3,x = 10:14, .name_repair = 'minimal')
tibble::tibble(x = 1:3,x = 11:13, .name_repair = 'unique')
#> New names:
#> • `x` -> `x...1`
#> • `x` -> `x...2`
#> # A tibble: 3 × 2
#>   x...1 x...2
#>   <int> <int>
#> 1     1    11
#> 2     2    12
#> 3     3    13
entibble(x = 1:3,x = 11:13, .name_repair = 'unique')
#> New names:
#> • `x` -> `x...1`
#> • `x` -> `x...2`
#> # A tibble: 3 × 2
#>   x...1 x...2
#>   <int> <int>
#> 1     1    11
#> 2     2    12
#> 3     3    13
entibble(x = 1:3,x = 11:13, .name_repair = 'unique_quiet')
#> # A tibble: 3 × 2
#>   x...1 x...2
#>   <int> <int>
#> 1     1    11
#> 2     2    12
#> 3     3    13
entibble(x = 1:3,x = 11:13, .name_repair = ~scrub_tabnames(., sep = ' <-- ', quiet = TRUE))
#> # A tibble: 3 × 2
#>   `x (-- 1` `x (-- 2`
#>       <int>     <int>
#> 1         1        11
#> 2         2        12
#> 3         3        13
enlist(head(iris), head(mtcars))
#> $`head(iris)`
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa
#> 
#> $`head(mtcars)`
#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> Valiant           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1
#> 

tibble::tibble(x = 1:3,x = 11:13, .name_repair = ~scrub_tabnames(., sep = ' <-- ', quiet = TRUE))
#> # A tibble: 3 × 2
#>   `x (-- 1` `x (-- 2`
#>       <int>     <int>
#> 1         1        11
#> 2         2        12
#> 3         3        13

as.matrix(1:10) |> rownames()
#> NULL
as.matrix(1:10) |> entibble()
#> New names:
#> • `` -> `...1`
#> # A tibble: 10 × 1
#>     ...1
#>    <int>
#>  1     1
#>  2     2
#>  3     3
#>  4     4
#>  5     5
#>  6     6
#>  7     7
#>  8     8
#>  9     9
#> 10    10
as.matrix(list(1:10,11:20)) |> entibble()
#> # A tibble: 10 × 2
#>      `1`   `2`
#>    <int> <int>
#>  1     1    11
#>  2     2    12
#>  3     3    13
#>  4     4    14
#>  5     5    15
#>  6     6    16
#>  7     7    17
#>  8     8    18
#>  9     9    19
#> 10    10    20
as.matrix(list(1:10,11:20)) |> tibble::tibble()
#> # A tibble: 2 × 1
#>   `as.matrix(list(1:10, 11:20))`[,1]
#>   <list>                            
#> 1 <int [10]>                        
#> 2 <int [10]>                        
as.matrix(1:10) |> tibble::tibble()
#> # A tibble: 10 × 1
#>    `as.matrix(1:10)`[,1]
#>                    <int>
#>  1                     1
#>  2                     2
#>  3                     3
#>  4                     4
#>  5                     5
#>  6                     6
#>  7                     7
#>  8                     8
#>  9                     9
#> 10                    10
as.matrix(warpbreaks[1:10, ]) |> tibble::has_rownames()
#> [1] FALSE
as.matrix(warpbreaks[1:10, ]) |> entibble()
#> # A tibble: 10 × 3
#>    breaks wool  tension
#>    <chr>  <chr> <chr>  
#>  1 26     A     L      
#>  2 30     A     L      
#>  3 54     A     L      
#>  4 25     A     L      
#>  5 70     A     L      
#>  6 52     A     L      
#>  7 51     A     L      
#>  8 26     A     L      
#>  9 67     A     L      
#> 10 18     A     M      
# ignores unused rownames
as.matrix(warpbreaks[1:10, ]) |> entibble(.rowname = 'row_nums')
#> # A tibble: 10 × 3
#>    breaks wool  tension
#>    <chr>  <chr> <chr>  
#>  1 26     A     L      
#>  2 30     A     L      
#>  3 54     A     L      
#>  4 25     A     L      
#>  5 70     A     L      
#>  6 52     A     L      
#>  7 51     A     L      
#>  8 26     A     L      
#>  9 67     A     L      
#> 10 18     A     M      
matrix(c(1, 2, 3, 11, 12, 13), nrow = 2, ncol = 3, byrow = TRUE, dimnames = list(c('row1', 'row2'), c('C.1', 'C.2', 'C.3'))) |> entibble()
#> # A tibble: 2 × 3
#>     C.1   C.2   C.3
#>   <dbl> <dbl> <dbl>
#> 1     1     2     3
#> 2    11    12    13
tail(iris) |> tibble::has_rownames()
#> [1] TRUE
tail(iris) |> entibble()
#> # A tibble: 6 × 6
#>   rowname Sepal.Length Sepal.Width Petal.Length Petal.Width Species  
#>   <chr>   <chr>        <chr>       <chr>        <chr>       <chr>    
#> 1 145     6.7          3.3         5.7          2.5         virginica
#> 2 146     6.7          3.0         5.2          2.3         virginica
#> 3 147     6.3          2.5         5.0          1.9         virginica
#> 4 148     6.5          3.0         5.2          2.0         virginica
#> 5 149     6.2          3.4         5.4          2.3         virginica
#> 6 150     5.9          3.0         5.1          1.8         virginica
mtcars |> entibble()
#> # A tibble: 32 × 12
#>    rowname       mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <chr>       <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Mazda RX4    21       6  160    110  3.9   2.62  16.5     0     1     4     4
#>  2 Mazda RX4 …  21       6  160    110  3.9   2.88  17.0     0     1     4     4
#>  3 Datsun 710   22.8     4  108     93  3.85  2.32  18.6     1     1     4     1
#>  4 Hornet 4 D…  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#>  5 Hornet Spo…  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2
#>  6 Valiant      18.1     6  225    105  2.76  3.46  20.2     1     0     3     1
#>  7 Duster 360   14.3     8  360    245  3.21  3.57  15.8     0     0     3     4
#>  8 Merc 240D    24.4     4  147.    62  3.69  3.19  20       1     0     4     2
#>  9 Merc 230     22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2
#> 10 Merc 280     19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4
#> # ℹ 22 more rows
mtcars |> entibble(.rowname = 'vehicle')
#> # A tibble: 32 × 12
#>    vehicle       mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <chr>       <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Mazda RX4    21       6  160    110  3.9   2.62  16.5     0     1     4     4
#>  2 Mazda RX4 …  21       6  160    110  3.9   2.88  17.0     0     1     4     4
#>  3 Datsun 710   22.8     4  108     93  3.85  2.32  18.6     1     1     4     1
#>  4 Hornet 4 D…  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#>  5 Hornet Spo…  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2
#>  6 Valiant      18.1     6  225    105  2.76  3.46  20.2     1     0     3     1
#>  7 Duster 360   14.3     8  360    245  3.21  3.57  15.8     0     0     3     4
#>  8 Merc 240D    24.4     4  147.    62  3.69  3.19  20       1     0     4     2
#>  9 Merc 230     22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2
#> 10 Merc 280     19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4
#> # ℹ 22 more rows
5 |> entibble()
#> # A tibble: 1 × 1
#>     `5`
#>   <dbl>
#> 1     5
NA |> entibble()
#> # A tibble: 1 × 1
#>   `NA` 
#>   <lgl>
#> 1 NA   
NULL |> entibble()
#> # A tibble: 0 × 0
enlist(head(iris), head(mtcars)) |> purrr::map(entibble)
#> $`head(iris)`
#> # A tibble: 6 × 5
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#>   <chr>        <chr>       <chr>        <chr>       <chr>  
#> 1 5.1          3.5         1.4          0.2         setosa 
#> 2 4.9          3.0         1.4          0.2         setosa 
#> 3 4.7          3.2         1.3          0.2         setosa 
#> 4 4.6          3.1         1.5          0.2         setosa 
#> 5 5.0          3.6         1.4          0.2         setosa 
#> 6 5.4          3.9         1.7          0.4         setosa 
#> 
#> $`head(mtcars)`
#> # A tibble: 6 × 12
#>   rowname        mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>   <chr>        <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Mazda RX4     21       6   160   110  3.9   2.62  16.5     0     1     4     4
#> 2 Mazda RX4 W…  21       6   160   110  3.9   2.88  17.0     0     1     4     4
#> 3 Datsun 710    22.8     4   108    93  3.85  2.32  18.6     1     1     4     1
#> 4 Hornet 4 Dr…  21.4     6   258   110  3.08  3.22  19.4     1     0     3     1
#> 5 Hornet Spor…  18.7     8   360   175  3.15  3.44  17.0     0     0     3     2
#> 6 Valiant       18.1     6   225   105  2.76  3.46  20.2     1     0     3     1
#> 
enlist(letters, LETTERS) |> purrr::map(entibble)
#> $letters
#> # A tibble: 26 × 1
#>    `<chr>`
#>    <chr>  
#>  1 a      
#>  2 b      
#>  3 c      
#>  4 d      
#>  5 e      
#>  6 f      
#>  7 g      
#>  8 h      
#>  9 i      
#> 10 j      
#> # ℹ 16 more rows
#> 
#> $LETTERS
#> # A tibble: 26 × 1
#>    `<chr>`
#>    <chr>  
#>  1 A      
#>  2 B      
#>  3 C      
#>  4 D      
#>  5 E      
#>  6 F      
#>  7 G      
#>  8 H      
#>  9 I      
#> 10 J      
#> # ℹ 16 more rows
#> 
enlist(letters, LETTERS) |> purrr::map(entibble, .name_repair = ~'a-z')
#> $letters
#> # A tibble: 26 × 1
#>    `a-z`
#>    <chr>
#>  1 a    
#>  2 b    
#>  3 c    
#>  4 d    
#>  5 e    
#>  6 f    
#>  7 g    
#>  8 h    
#>  9 i    
#> 10 j    
#> # ℹ 16 more rows
#> 
#> $LETTERS
#> # A tibble: 26 × 1
#>    `a-z`
#>    <chr>
#>  1 A    
#>  2 B    
#>  3 C    
#>  4 D    
#>  5 E    
#>  6 F    
#>  7 G    
#>  8 H    
#>  9 I    
#> 10 J    
#> # ℹ 16 more rows
#> 
enlist(head(iris), head(mtcars)) |> entibble()
#> # A tibble: 6 × 17
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species rowname    mpg   cyl
#>   <chr>        <chr>       <chr>        <chr>       <chr>   <chr>    <dbl> <dbl>
#> 1 5.1          3.5         1.4          0.2         setosa  Mazda R…  21       6
#> 2 4.9          3.0         1.4          0.2         setosa  Mazda R…  21       6
#> 3 4.7          3.2         1.3          0.2         setosa  Datsun …  22.8     4
#> 4 4.6          3.1         1.5          0.2         setosa  Hornet …  21.4     6
#> 5 5.0          3.6         1.4          0.2         setosa  Hornet …  18.7     8
#> 6 5.4          3.9         1.7          0.4         setosa  Valiant   18.1     6
#> # ℹ 9 more variables: disp <dbl>, hp <dbl>, drat <dbl>, wt <dbl>, qsec <dbl>,
#> #   vs <dbl>, am <dbl>, gear <dbl>, carb <dbl>
head(iris) |> purrr::map(entibble)
#> $Sepal.Length
#> # A tibble: 6 × 1
#>   `<dbl>`
#>     <dbl>
#> 1     5.1
#> 2     4.9
#> 3     4.7
#> 4     4.6
#> 5     5  
#> 6     5.4
#> 
#> $Sepal.Width
#> # A tibble: 6 × 1
#>   `<dbl>`
#>     <dbl>
#> 1     3.5
#> 2     3  
#> 3     3.2
#> 4     3.1
#> 5     3.6
#> 6     3.9
#> 
#> $Petal.Length
#> # A tibble: 6 × 1
#>   `<dbl>`
#>     <dbl>
#> 1     1.4
#> 2     1.4
#> 3     1.3
#> 4     1.5
#> 5     1.4
#> 6     1.7
#> 
#> $Petal.Width
#> # A tibble: 6 × 1
#>   `<dbl>`
#>     <dbl>
#> 1     0.2
#> 2     0.2
#> 3     0.2
#> 4     0.2
#> 5     0.2
#> 6     0.4
#> 
#> $Species
#> # A tibble: 6 × 1
#>   `<fct>`
#>   <fct>  
#> 1 setosa 
#> 2 setosa 
#> 3 setosa 
#> 4 setosa 
#> 5 setosa 
#> 6 setosa 
#> 
entibble(!!!head(iris))
#> # A tibble: 6 × 5
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#>          <dbl>       <dbl>        <dbl>       <dbl> <fct>  
#> 1          5.1         3.5          1.4         0.2 setosa 
#> 2          4.9         3            1.4         0.2 setosa 
#> 3          4.7         3.2          1.3         0.2 setosa 
#> 4          4.6         3.1          1.5         0.2 setosa 
#> 5          5           3.6          1.4         0.2 setosa 
#> 6          5.4         3.9          1.7         0.4 setosa 
enlist(head(iris), head(mtcars)) |> entibble()
#> # A tibble: 6 × 17
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species rowname    mpg   cyl
#>   <chr>        <chr>       <chr>        <chr>       <chr>   <chr>    <dbl> <dbl>
#> 1 5.1          3.5         1.4          0.2         setosa  Mazda R…  21       6
#> 2 4.9          3.0         1.4          0.2         setosa  Mazda R…  21       6
#> 3 4.7          3.2         1.3          0.2         setosa  Datsun …  22.8     4
#> 4 4.6          3.1         1.5          0.2         setosa  Hornet …  21.4     6
#> 5 5.0          3.6         1.4          0.2         setosa  Hornet …  18.7     8
#> 6 5.4          3.9         1.7          0.4         setosa  Valiant   18.1     6
#> # ℹ 9 more variables: disp <dbl>, hp <dbl>, drat <dbl>, wt <dbl>, qsec <dbl>,
#> #   vs <dbl>, am <dbl>, gear <dbl>, carb <dbl>
enlist(head(iris), head(mtcars)) |> tibble::tibble()
#> # A tibble: 2 × 1
#>   `enlist(head(iris), head(mtcars))`
#>   <named list>                      
#> 1 <df [6 × 5]>                      
#> 2 <df [6 × 11]>                     

letters |> purrr::set_names(LETTERS) |> tibble::tibble()
#> # A tibble: 26 × 1
#>    `purrr::set_names(letters, LETTERS)`
#>    <chr>                               
#>  1 a                                   
#>  2 b                                   
#>  3 c                                   
#>  4 d                                   
#>  5 e                                   
#>  6 f                                   
#>  7 g                                   
#>  8 h                                   
#>  9 i                                   
#> 10 j                                   
#> # ℹ 16 more rows
letters |> purrr::set_names(LETTERS) |> entibble()
#> # A tibble: 26 × 2
#>    rowname `purrr::set_names(letters, LETTERS)`
#>    <chr>   <chr>                               
#>  1 A       a                                   
#>  2 B       b                                   
#>  3 C       c                                   
#>  4 D       d                                   
#>  5 E       e                                   
#>  6 F       f                                   
#>  7 G       g                                   
#>  8 H       h                                   
#>  9 I       i                                   
#> 10 J       j                                   
#> # ℹ 16 more rows

letters |> purrr::set_names(LETTERS) |> tibble::enframe( name = 'row', value = 'blue')
#> # A tibble: 26 × 2
#>    row   blue 
#>    <chr> <chr>
#>  1 A     a    
#>  2 B     b    
#>  3 C     c    
#>  4 D     d    
#>  5 E     e    
#>  6 F     f    
#>  7 G     g    
#>  8 H     h    
#>  9 I     i    
#> 10 J     j    
#> # ℹ 16 more rows

# If rowname is specified, but the dataframe has no rownames, rowname is ignored.
entibble(iris, .rowname = 'flower_name')
#> # A tibble: 150 × 5
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#>    <chr>        <chr>       <chr>        <chr>       <chr>  
#>  1 5.1          3.5         1.4          0.2         setosa 
#>  2 4.9          3.0         1.4          0.2         setosa 
#>  3 4.7          3.2         1.3          0.2         setosa 
#>  4 4.6          3.1         1.5          0.2         setosa 
#>  5 5.0          3.6         1.4          0.2         setosa 
#>  6 5.4          3.9         1.7          0.4         setosa 
#>  7 4.6          3.4         1.4          0.3         setosa 
#>  8 5.0          3.4         1.5          0.2         setosa 
#>  9 4.4          2.9         1.4          0.2         setosa 
#> 10 4.9          3.1         1.5          0.1         setosa 
#> # ℹ 140 more rows
letters |> as.list() |> tibble::tibble()
#> # A tibble: 26 × 1
#>    `as.list(letters)`
#>    <list>            
#>  1 <chr [1]>         
#>  2 <chr [1]>         
#>  3 <chr [1]>         
#>  4 <chr [1]>         
#>  5 <chr [1]>         
#>  6 <chr [1]>         
#>  7 <chr [1]>         
#>  8 <chr [1]>         
#>  9 <chr [1]>         
#> 10 <chr [1]>         
#> # ℹ 16 more rows
letters |> as.list() |> entibble(.rowname = 'a-z')
#> # A tibble: 1 × 26
#>   `1`   `2`   `3`   `4`   `5`   `6`   `7`   `8`   `9`   `10`  `11`  `12`  `13` 
#>   <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#> 1 a     b     c     d     e     f     g     h     i     j     k     l     m    
#> # ℹ 13 more variables: `14` <chr>, `15` <chr>, `16` <chr>, `17` <chr>,
#> #   `18` <chr>, `19` <chr>, `20` <chr>, `21` <chr>, `22` <chr>, `23` <chr>,
#> #   `24` <chr>, `25` <chr>, `26` <chr>
letters |> list() |> tibble::tibble()
#> # A tibble: 1 × 1
#>   `list(letters)`
#>   <list>         
#> 1 <chr [26]>     
letters |> list() |> entibble(.rowname = 'a-z')
#> # A tibble: 26 × 1
#>    `list(letters)`
#>    <chr>          
#>  1 a              
#>  2 b              
#>  3 c              
#>  4 d              
#>  5 e              
#>  6 f              
#>  7 g              
#>  8 h              
#>  9 i              
#> 10 j              
#> # ℹ 16 more rows
letters |> enlist() |> entibble(.rowname = 'a-z')
#> # A tibble: 26 × 1
#>    `enlist(letters)`
#>    <chr>            
#>  1 a                
#>  2 b                
#>  3 c                
#>  4 d                
#>  5 e                
#>  6 f                
#>  7 g                
#>  8 h                
#>  9 i                
#> 10 j                
#> # ℹ 16 more rows

colours() |> rlang::set_names() |> entibble() |> entibble(another_col = colours())
#> # A tibble: 657 × 3
#>    rowname       `rlang::set_names(colours())` another_col  
#>    <chr>         <chr>                         <chr>        
#>  1 white         white                         white        
#>  2 aliceblue     aliceblue                     aliceblue    
#>  3 antiquewhite  antiquewhite                  antiquewhite 
#>  4 antiquewhite1 antiquewhite1                 antiquewhite1
#>  5 antiquewhite2 antiquewhite2                 antiquewhite2
#>  6 antiquewhite3 antiquewhite3                 antiquewhite3
#>  7 antiquewhite4 antiquewhite4                 antiquewhite4
#>  8 aquamarine    aquamarine                    aquamarine   
#>  9 aquamarine1   aquamarine1                   aquamarine1  
#> 10 aquamarine2   aquamarine2                   aquamarine2  
#> # ℹ 647 more rows
colours() |> rlang::set_names() |> entibble(another_col = colours())
#> # A tibble: 657 × 3
#>    rowname       `rlang::set_names(colours())` another_col  
#>    <chr>         <chr>                         <chr>        
#>  1 white         white                         white        
#>  2 aliceblue     aliceblue                     aliceblue    
#>  3 antiquewhite  antiquewhite                  antiquewhite 
#>  4 antiquewhite1 antiquewhite1                 antiquewhite1
#>  5 antiquewhite2 antiquewhite2                 antiquewhite2
#>  6 antiquewhite3 antiquewhite3                 antiquewhite3
#>  7 antiquewhite4 antiquewhite4                 antiquewhite4
#>  8 aquamarine    aquamarine                    aquamarine   
#>  9 aquamarine1   aquamarine1                   aquamarine1  
#> 10 aquamarine2   aquamarine2                   aquamarine2  
#> # ℹ 647 more rows
letters |> as.list() |> enlist() |> entibble()
#> # A tibble: 1 × 26
#>   `1`   `2`   `3`   `4`   `5`   `6`   `7`   `8`   `9`   `10`  `11`  `12`  `13` 
#>   <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#> 1 a     b     c     d     e     f     g     h     i     j     k     l     m    
#> # ℹ 13 more variables: `14` <chr>, `15` <chr>, `16` <chr>, `17` <chr>,
#> #   `18` <chr>, `19` <chr>, `20` <chr>, `21` <chr>, `22` <chr>, `23` <chr>,
#> #   `24` <chr>, `25` <chr>, `26` <chr>
# dplyr::bind_cols(letters, LETTERS) |> copy_for_xl()
# letters |> rlang::set_names(LETTERS) |> copy_for_xl()

entibble(1:10,11:20)
#> # A tibble: 10 × 2
#>    `1:10` `11:20`
#>     <int>   <int>
#>  1      1      11
#>  2      2      12
#>  3      3      13
#>  4      4      14
#>  5      5      15
#>  6      6      16
#>  7      7      17
#>  8      8      18
#>  9      9      19
#> 10     10      20
entibble(c(1:10,11:20))
#> # A tibble: 20 × 1
#>    `c(1:10, 11:20)`
#>               <int>
#>  1                1
#>  2                2
#>  3                3
#>  4                4
#>  5                5
#>  6                6
#>  7                7
#>  8                8
#>  9                9
#> 10               10
#> 11               11
#> 12               12
#> 13               13
#> 14               14
#> 15               15
#> 16               16
#> 17               17
#> 18               18
#> 19               19
#> 20               20

entibble(list(1:10,11:20))
#> # A tibble: 10 × 2
#>    `1:10` `11:20`
#>     <int>   <int>
#>  1      1      11
#>  2      2      12
#>  3      3      13
#>  4      4      14
#>  5      5      15
#>  6      6      16
#>  7      7      17
#>  8      8      18
#>  9      9      19
#> 10     10      20
tibble::tibble(list(1:10,11:20))
#> # A tibble: 2 × 1
#>   `list(1:10, 11:20)`
#>   <list>             
#> 1 <int [10]>         
#> 2 <int [10]>         

entibble(enlist(1:10,11:20))
#> # A tibble: 10 × 2
#>    `1:10` `11:20`
#>     <int>   <int>
#>  1      1      11
#>  2      2      12
#>  3      3      13
#>  4      4      14
#>  5      5      15
#>  6      6      16
#>  7      7      17
#>  8      8      18
#>  9      9      19
#> 10     10      20
entibble(enlist(1:10,11:20), .rowname = 'nums')
#> # A tibble: 10 × 2
#>    `1:10` `11:20`
#>     <int>   <int>
#>  1      1      11
#>  2      2      12
#>  3      3      13
#>  4      4      14
#>  5      5      15
#>  6      6      16
#>  7      7      17
#>  8      8      18
#>  9      9      19
#> 10     10      20
entibble(list(1:10),list(11:20))
#> # A tibble: 1 × 2
#>   `list(1:10)` `list(11:20)`
#>   <list>       <list>       
#> 1 <int [10]>   <int [10]>   
tibble::tibble(enlist(1:10),enlist(11:20))
#> # A tibble: 1 × 2
#>   `enlist(1:10)` `enlist(11:20)`
#>   <named list>   <named list>   
#> 1 <int [10]>     <int [10]>     
entibble(enlist(1:10),enlist(11:20))
#> # A tibble: 1 × 4
#>   rowname `enlist(1:10)` rowname `enlist(11:20)`
#>   <chr>   <list>         <chr>   <list>         
#> 1 1:10    <int [10]>     11:20   <int [10]>     
entibble(enlist(1:10),enlist(11:20), .rowname = 'blue')
#> # A tibble: 1 × 4
#>   blue  `enlist(1:10)` blue  `enlist(11:20)`
#>   <chr> <list>         <chr> <list>         
#> 1 1:10  <int [10]>     11:20 <int [10]>     
entibble(enlist(1:10),enlist(11:20), .rowname = 'a column name', .name_repair = 'unique')
#> New names:
#> • `a column name` -> `a column name...1`
#> • `a column name` -> `a column name...3`
#> # A tibble: 1 × 4
#>   `a column name...1` `enlist(1:10)` `a column name...3` `enlist(11:20)`
#>   <chr>               <list>         <chr>               <list>         
#> 1 1:10                <int [10]>     11:20               <int [10]>     
entibble(enlist(1:10),enlist(11:20), .rowname = 'a column name', .name_repair = 'universal')
#> New names:
#> • `a column name` -> `a.column.name...1`
#> • `enlist(1:10)` -> `enlist.1.10.`
#> • `a column name` -> `a.column.name...3`
#> • `enlist(11:20)` -> `enlist.11.20.`
#> # A tibble: 1 × 4
#>   a.column.name...1 enlist.1.10. a.column.name...3 enlist.11.20.
#>   <chr>             <list>       <chr>             <list>       
#> 1 1:10              <int [10]>   11:20             <int [10]>   
entibble(enlist(1:10),enlist(11:20), .rowname = 'a column name', .name_repair = 'universal_quiet')
#> # A tibble: 1 × 4
#>   a.column.name...1 enlist.1.10. a.column.name...3 enlist.11.20.
#>   <chr>             <list>       <chr>             <list>       
#> 1 1:10              <int [10]>   11:20             <int [10]>   
entibble(enlist(list(1:10),list(11:20)))
#> # A tibble: 1 × 2
#>   `list(1:10)` `list(11:20)`
#>   <list>       <list>       
#> 1 <int [10]>   <int [10]>   
entibble(enlist((1:10),(11:20)))
#> # A tibble: 10 × 2
#>    `(1:10)` `(11:20)`
#>       <int>     <int>
#>  1        1        11
#>  2        2        12
#>  3        3        13
#>  4        4        14
#>  5        5        15
#>  6        6        16
#>  7        7        17
#>  8        8        18
#>  9        9        19
#> 10       10        20